+++ lib/features/app/presentation/app_bloc/app_bloc.dart
@@ -6,7 +6,6 @@ import 'package:logging/logging.dart';
 import 'package:malawi_ride_share_app/features/app/domain/usecases/ensure_location_permission.dart';
 import 'package:malawi_ride_share_app/features/app/domain/usecases/ensure_notification_permission.dart';
 import 'package:malawi_ride_share_app/features/app/domain/usecases/open_location_settings.dart';
-import 'package:malawi_ride_share_app/features/app/data/repositories/firebase_repository.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 part 'app_event.dart';
 part 'app_state.dart';
