+++ lib/features/driver/domain/repository/driver_life_cycle_management.dart
@@ -1,5 +1,5 @@
 abstract class DriverLifeCycleManagement {
-  Future<void> initializeDriver({required String firebaseId});
+  Future<void> initializeDriverSession({required String firebaseId});
   Future<void> goOnline({required String firebaseId});
   Future<void> goOffline({required String firebaseId});
   Future<void> updateDriverAvalability();
