+++ lib/features/shared/domain/repositories/firebase_repository.dart
@@ -4,5 +4,6 @@ abstract class FirebaseRepository {
   Future<String?> initializeMessaging();
   Future<String?> getFCMToken();
   Future<User> getCurrentUser();
+  String? getFirebaseId();
   Future<void> registerDevice();
 }
